+++ glfw/wl_window.c
@@ -451,11 +451,11 @@ static void xdgToplevelHandleConfigure(void* data,
     debug("final window content size: %dx%d\n", window->wl.width, window->wl.height);
     _glfwInputWindowFocus(window, window->wl.toplevel_states & TOPLEVEL_STATE_ACTIVATED);
     ensure_csd_resources(window);
+    wl_surface_commit(window->wl.surface);
 #define geometry window->wl.decorations.geometry
     debug("Setting window geometry: x=%d y=%d %dx%d\n", geometry.x, geometry.y, geometry.width, geometry.height);
     xdg_surface_set_window_geometry(window->wl.xdg.surface, geometry.x, geometry.y, geometry.width, geometry.height);
 #undef geometry
-    wl_surface_commit(window->wl.surface);
     if (live_resize_done) _glfwInputLiveResize(window, false);
 }