+++ docs/changelog.rst
@@ -16,6 +16,10 @@ To update |kitty|, :doc:`follow the instructions <binary>`.
 - Fix ``kitty @ close-tab`` not working with pipe based remote control
   (:iss:`3510`)
 
+- Fix removal of inactive tab that is before the currently active tab causing
+  the highlighted tab to be incorrect (:iss:`3516`)
+
+
 0.20.1 [2021-04-19]
 ----------------------